(feat)ObtenerDetallePorIdAsync en Trabajo
Al llamar desde el endpoint para obtener el trabajo detallado se completan todas las relaciones

+++ ShopMGR.Dominio/Modelo/Material.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Descripcion { get; set; }
         public decimal Precio { get; set; }
-        public float Cantidad { get; set; }
+        public double Cantidad { get; set; }
 
         //Relaciones
         public Presupuesto Presupuesto { get; set; }
